ARC: Disallow RTSC for SMP configs
ARC700 RTSC has issues and it can't be used reliably. The only reason it
was kept around was a customer using it to implement cross-core
synchronized timer (as a truely exception use case since this thing
is strictly in-core).

Anyhow that usage is no longer supported so we can remove RTSC
altogether. However ARCv2 ISA has a saner implementation of 64bit timer
- with different mechanics ofcourse. To reduce the code churn we keep
the dead code as of now and when time comes simply change the timer
backend with the existing Kconfig infrastructure.

For now we simply remove the special SMP glue.

Signed-off-by: Vineet Gupta <vgupta@synopsys.com>