ARC: [SMP] send ipi to set ienable.
In case of not percpu interrupt can force ienable mask be symetric
across CPUs. This solves problem when peripheral is not
connected to all CPUs and tries to interrupt masked CPU.

It may be optimized if we got prior knowledge which cpu needs this IPI
instead brodcasting all CPUs.

Signed-off-by: Noam Camus <noamc@ezchip.com>
Signed-off-by: Vineet Gupta <vgupta@synopsys.com>

Conflicts:
	arch/arc/kernel/smp.c